refactor: move types to lib/types directory
